build: fix mechanism for additional dependencies
Instead of expecting plugins to store their configuration files in
a "configs/" directory, read the dependencies from the config.xml file.
The reason why the old method does not work is that the <resource-file>
element of the plugin.xml file doesn't support a "target-dir" attribute.

By using the <config-file> element, instead, we don't need to do any
changes to cordova lib, and all the plugin configuration can now be
stored in the plugin.xml file.

Example:

<config-file target="config.xml" parent="/*">
  <feature name="FacebookConnect">
    <param policy_group="accounts" policy_version="1" />
    <param hook="account-application" value="qml/app.application" />
    <param hook="account-service" value="qml/app.service" />
    <deps>
      <pkgConfig>["signond", "libsignon-qt5"]</pkgConfig>
      <deb>["signond-dev:ARCH", "libsignon-qt5-dev:ARCH"]</deb>
    </deps>
  </feature>
</config-file>